Create config package
